LPS-62388 Renaming classes for more clarity
