+++ modules/apps/export-import/export-import-web/src/main/java/com/liferay/exportimport/web/portlet/action/DeleteLayoutImportBackgroundTaskMVCActionCommand.java
@@ -15,11 +15,13 @@
 package com.liferay.exportimport.web.portlet.action;
 
 import com.liferay.exportimport.web.constants.ExportImportPortletKeys;
+import com.liferay.portal.kernel.backgroundtask.BackgroundTaskManagerUtil;
 import com.liferay.portal.kernel.exception.NoSuchBackgroundTaskException;
 import com.liferay.portal.kernel.portlet.bridges.mvc.BaseMVCActionCommand;
 import com.liferay.portal.kernel.portlet.bridges.mvc.MVCActionCommand;
 import com.liferay.portal.kernel.security.auth.PrincipalException;
 import com.liferay.portal.kernel.servlet.SessionErrors;
+import com.liferay.portal.kernel.util.ParamUtil;
 
 import javax.portlet.ActionRequest;
 import javax.portlet.ActionResponse;
@@ -32,12 +34,13 @@ import org.osgi.service.component.annotations.Component;
 @Component(
 	immediate = true,
 	property = {
-		"javax.portlet.name=" + ExportImportPortletKeys.EXPORT_IMPORT,
-		"mvc.command.name=deleteBackgroundTask"
+		"javax.portlet.name=" + ExportImportPortletKeys.IMPORT,
+		"mvc.command.name=deleteBackgroundTasks"
 	},
 	service = MVCActionCommand.class
 )
-public class DeleteBackgroundTaskMVCActionCommand extends BaseMVCActionCommand {
+public class DeleteLayoutImportBackgroundTaskMVCActionCommand
+	extends BaseMVCActionCommand {
 
 	@Override
 	protected void doProcessAction(
@@ -45,7 +48,13 @@ public class DeleteBackgroundTaskMVCActionCommand extends BaseMVCActionCommand {
 		throws Exception {
 
 		try {
-			ActionUtil.deleteBackgroundTask(actionRequest);
+			long[] backgroundTaskIds = ParamUtil.getLongValues(
+				actionRequest, "deleteBackgroundTaskIds");
+
+			for (long backgroundTaskId : backgroundTaskIds) {
+				BackgroundTaskManagerUtil.deleteBackgroundTask(
+					backgroundTaskId);
+			}
 		}
 		catch (Exception e) {
 			if (e instanceof NoSuchBackgroundTaskException ||
