+++ modules/apps/staging/staging-processes-web/src/main/java/com/liferay/staging/processes/web/portlet/action/DeleteLayoutPublishBackgroundTaskMVCActionCommand.java
@@ -12,16 +12,17 @@
  * details.
  */
 
-package com.liferay.exportimport.web.portlet.action;
+package com.liferay.staging.processes.web.portlet.action;
 
-import com.liferay.exportimport.web.constants.ExportImportPortletKeys;
 import com.liferay.portal.kernel.backgroundtask.BackgroundTaskManagerUtil;
 import com.liferay.portal.kernel.exception.NoSuchBackgroundTaskException;
+import com.liferay.portal.kernel.exception.PortalException;
 import com.liferay.portal.kernel.portlet.bridges.mvc.BaseMVCActionCommand;
 import com.liferay.portal.kernel.portlet.bridges.mvc.MVCActionCommand;
 import com.liferay.portal.kernel.security.auth.PrincipalException;
 import com.liferay.portal.kernel.servlet.SessionErrors;
 import com.liferay.portal.kernel.util.ParamUtil;
+import com.liferay.staging.constants.StagingProcessesPortletKeys;
 
 import javax.portlet.ActionRequest;
 import javax.portlet.ActionResponse;
@@ -29,18 +30,29 @@ import javax.portlet.ActionResponse;
 import org.osgi.service.component.annotations.Component;
 
 /**
- * @author Daniel Kocsis
+ * @author Levente Hudák
  */
 @Component(
 	immediate = true,
 	property = {
-		"javax.portlet.name=" + ExportImportPortletKeys.EXPORT,
+		"javax.portlet.name=" + StagingProcessesPortletKeys.STAGING_PROCESSES,
 		"mvc.command.name=deleteBackgroundTasks"
 	},
 	service = MVCActionCommand.class
 )
-public class DeleteExportBackgroundTaskMVCActionCommand
+public class DeleteLayoutPublishBackgroundTaskMVCActionCommand
 	extends BaseMVCActionCommand {
+
+	protected void deleteBackgroundTask(ActionRequest actionRequest)
+		throws PortalException {
+
+		long[] backgroundTaskIds = ParamUtil.getLongValues(
+			actionRequest, "deleteBackgroundTaskIds");
+
+		for (long backgroundTaskId : backgroundTaskIds) {
+			BackgroundTaskManagerUtil.deleteBackgroundTask(backgroundTaskId);
+		}
+	}
 
 	@Override
 	protected void doProcessAction(
@@ -48,13 +60,7 @@ public class DeleteExportBackgroundTaskMVCActionCommand
 		throws Exception {
 
 		try {
-			long[] backgroundTaskIds = ParamUtil.getLongValues(
-				actionRequest, "deleteBackgroundTaskIds");
-
-			for (long backgroundTaskId : backgroundTaskIds) {
-				BackgroundTaskManagerUtil.deleteBackgroundTask(
-					backgroundTaskId);
-			}
+			deleteBackgroundTask(actionRequest);
 		}
 		catch (Exception e) {
 			if (e instanceof NoSuchBackgroundTaskException ||
@@ -62,7 +68,8 @@ public class DeleteExportBackgroundTaskMVCActionCommand
 
 				SessionErrors.add(actionRequest, e.getClass());
 
-				actionResponse.setRenderParameter("mvcPath", "/error.jsp");
+				actionResponse.setRenderParameter(
+					"mvcPath", "/error/error.jsp");
 			}
 			else {
 				throw e;
